pyproject.toml
[project]
name = "table-dataset-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-functions>=1.24.0",
    "azure-functions-durable>=1.4.0",
    "azure-storage-blob>=12.27.0",
    "dotenv>=0.9.9",
    "google>=3.0.0",
    "google-api-python-client>=2.185.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "ipykernel>=7.0.1",
    "playwright>=1.55.0",
    "pydantic>=2.12.3",
]
